Feat: Add directory listing function
This commit adds a directory listing function to the HTTP server.
With this function, users can get a list of content (including
subdirectories and files) in the directory they specify by URL path.
The server can now return the content list if the URL path is
a directory, or return the file content if it is a file.

To implement the directory listing function, a new argument
to the module WWWROOT has been added. This argument specifies
the HTTP server root directory. Users' URL paths must be based on
this root path.

In the http_server_response() function, a new function called
directory_handler(struct http_request, int) has been added.
This function takes the requested path and opens the specified directory
(or file) and checks its type using the S_ISDIR and S_ISREG macros.
If it is a directory, the server iterates through the directory using
iterate_dir() and returns the content list in HTML with
'Transfer-Encoding: chunked'. If it is a file, the server returns
the file content in HTML with the connection type specified by the user.

This commit enhances the functionality of the HTTP server and improves
the user experience by providing directory listing capability.