Remove CMWQ work from http_request
Create a new structure called 'struct httpd_work'
that includes a member of type 'work_struct',
instead of adding 'work_struct' as a member
of  'struct http_request'.

The purpose of attaching 'struct work' to
'struct request' is solely to locate the
corresponding socket, and the remaining
members of 'struct request' are not used.
Therefore, mounting it on 'struct httpd_work'
instead can save some space.