Feat: Add content cache
Add content caching to improve server
throughput. When users frequently access
certain URLs, it can significantly speed up
webpage responses. The data related to
accessed URLs will be stored in a hash table
(key: URL, value: corresponding content) in
memory, managed by Read-Copy Update
(RCU) to avoid the repetitive execution of
kernel_read() or iterate_dir(), which can be
time-consuming.

As RCU is used, there is no locking during
cache data reading. Besides, implementing
a timer to periodically clear expired cache
content to avoid excessive memory usage.

Throughput without content cache
requests/sec:  3346.166

Throughput with content cache
requests/sec:  16351.585

Please note the potential limitations of
this caching design:

The cache capacity is limited. For file
content, the available memory depends on the
maximum limit of kmalloc(), which may vary on
different devices. For file directories, an
array of 20,000 bytes is used for storage. If
these capacity limits are exceeded, an error
may occur.

When a URL is accessed for the first time, its
content is cached, and the timer starts.
Subsequent accesses will retrieve the cached
content, but the timer will not be reset.
If the timer expires, the cached content
will be deleted and wait for the next
creation.